# Optimize `find_first_of` for one element needle