--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -3489,11 +3489,13 @@ namespace ranges {
             _STL_INTERNAL_STATIC_ASSERT(sentinel_for<_Se2, _It2>);
             _STL_INTERNAL_STATIC_ASSERT(indirectly_comparable<_It1, _It2, _Pr, _Pj1, _Pj2>);
 
-            if constexpr (_Is_ranges_random_iter_v<_It2> && sized_sentinel_for<_Se2, _It2>) {
+
+            if constexpr (_Is_ranges_random_iter_v<_It2> && sized_sentinel_for<_Se2, _It2>
+                          && _Is_any_of_v<_Pr, _STD equal_to<>, _RANGES equal_to>) {
                 const auto _Count2 = _Last2 - _First2;
                 if (_Count2 == 1) {
                     return _RANGES _Find_unchecked(
-                        _STD move(_First1), _STD move(_Last1), _STD invoke(_Proj2, *_First2), _Proj1);
+                        _STD move(_First1), _STD move(_Last1), _STD invoke(_Proj2, *_First2), _STD _Pass_fn(_Proj1));
                 }
             }