--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -3384,7 +3384,7 @@ _NODISCARD _CONSTEXPR20 _FwdIt1 find_first_of(
     const auto _UFirst2 = _STD _Get_unwrapped(_First2);
     const auto _ULast2  = _STD _Get_unwrapped(_Last2);
 
-    if constexpr (_Is_ranges_random_iter_v<_FwdIt2>) {
+    if constexpr (_Is_ranges_random_iter_v<_FwdIt2> && _Is_any_of_v<_Pr, _STD equal_to<>, _RANGES equal_to>) {
         const auto _Count2 = _Last2 - _First2;
         if (_Count2 == 1) {
             _UFirst1 = _STD _Find_unchecked(_STD move(_UFirst1), _STD move(_ULast1), *_UFirst2);