--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -3500,7 +3500,7 @@ namespace ranges {
                 const auto _Count2 = _Last2 - _First2;
                 if (_Count2 == 1) {
                     return _RANGES _Find_unchecked(
-                        _STD move(_First1), _STD move(_Last1), _STD invoke(_Proj2, *_First2), _STD _Pass_fn(_Proj1));
+                        _STD move(_First1), _STD move(_Last1), _STD invoke(_Proj2, *_First2), _Proj1);
                 }
             }